feat: recruiter assistant, shared embedding store, documents & model lab
Chatbot as a service (not a third agent): corpus-derived lexicon, typed query
understanding, staged pool retrieval, guardrails and conversation memory.

Shared embedding store replaces the two divergent indexes — resumes are parsed
and chunked once, then embedded per model over the same chunk set, so the three
engines are comparable and coverage gaps are visible instead of silent.

Adds the Documents working-set screen, Model Lab comparison, match-score
breakdown and the candidate portal views.

+++ ai_hiring_platform/backend/app/schemas/analytics.py
@@ -60,7 +60,7 @@ class RecentAnalysisItem(BaseModel):
     timestamp: datetime
     resume_filename: str
     jd_filename: str
-    overall_score: Optional[int] = None
+    overall_score: Optional[float] = None
     recommendation: str
     workflow_status: str
